[packages/template][s]: pass authorDetails to blog layout as prop
